There is much wrong with this code!
Instead of being tied to the set dimensions, this should *in theory* work for any sized source image, split into any number of tiles. However it totally doesn't when a whole load of edge cases.

Need to go back in and sort it out a bit more by dumping values down and studying them.

Anyway, in the meantime it still works for our very specific case, but really, don't use this latest commit ;)